Refactor storage strategies so buffer APIs use stream APIs
This means that subclasses now only have to implement the streaming version, rather than both streaming and buffer versions.
