+++ lib/io/storage/StorageStrategyBase.js
@@ -1,5 +1,10 @@
 /* eslint-disable class-methods-use-this, no-unused-vars */
 import { NotImplementedError } from '@/lib/error/MoveErrors';
+import {
+  bufferToDuplexStream,
+  readableStreamToBuffer,
+  writableStreamFinish,
+} from '@/lib/io/StreamUtils';
 
 /**
  * Base class for file storage strategies in MOVE.  In MOVE, a file storage strategy is
@@ -21,13 +26,13 @@ class StorageStrategyBase {
   }
 
   /**
-   * @abstract
    * @param {string} namespace
    * @param {string} key
-   * @returns {Promise<Buffer?>} value at the given `namespace` and `key`, or `null`
+   * @returns {Promise<Buffer>} value at the given `namespace` and `key`
    */
   async get(namespace, key) {
-    throw new NotImplementedError();
+    const valueStream = this.getStream(namespace, key);
+    return readableStreamToBuffer(valueStream);
   }
 
   /**
@@ -42,14 +47,15 @@ class StorageStrategyBase {
   }
 
   /**
-   * @abstract
    * @param {string} namespace
    * @param {string} key
    * @param {Buffer} value
-   * @returns {Promise<boolean>} whether the operation succeeded
+   * @returns {Promise<void>} when the value has finished writing
    */
   async put(namespace, key, value) {
-    throw new NotImplementedError();
+    const valueStream = bufferToDuplexStream(value);
+    const writableStream = await this.putStream(namespace, key, valueStream);
+    await writableStreamFinish(writableStream);
   }
 
   /**
