+++ lib/io/storage/StorageStrategyFilesystem.js
@@ -44,31 +44,9 @@ class StorageStrategyFilesystem extends StorageStrategyBase {
     }
   }
 
-  async get(namespace, key) {
-    const fullPath = this.getFullPath(namespace, key);
-    try {
-      const value = await fs.promises.readFile(fullPath);
-      return value;
-    } catch (err) {
-      return null;
-    }
-  }
-
   getStream(namespace, key) {
     const fullPath = this.getFullPath(namespace, key);
     return fs.createReadStream(fullPath);
-  }
-
-  async put(namespace, key, value) {
-    const fullPath = this.getFullPath(namespace, key);
-    const dir = path.dirname(fullPath);
-    await fs.promises.mkdir(dir, { recursive: true });
-    try {
-      await fs.promises.writeFile(fullPath, value);
-      return true;
-    } catch (err) {
-      return null;
-    }
   }
 
   async putStream(namespace, key, valueStream) {
